backend/requirements.txt
openai>=1.0.0
supabase>=2.0.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
fastapi>=0.109.0
uvicorn>=0.27.0
numpy<2
